🎨 Chore: InputComponent 분리
